Module riscv_5stage_simulator::hazards

Data hazard detection.

Functions

ex_hazard_src1
Indicates ALU src1 should be forwarded from the previous ALU result.

ex_hazard_src2
Indicates ALU src2 should be forwarded from the previous ALU result.

load_hazard
Indicates a load-use hazard that will require a pipeline stall.

mem_hazard_src1
Indicates ALU src1 should be forwarded from data memory or earlier result.

mem_hazard_src2
Indicates ALU src2 should be forwarded from data memory or earlier result.

reg_hazard_src1
Indicates src1 register was just written to and should be forwarded.

reg_hazard_src2
Indicates src2 register was just written to and should be forwarded.